pytorch_classifier/requirements/requirements.txt
cudatoolkit=11.0.221
pytorch=1.7.1
torchvision=0.8.2
pytorch-lightning=1.3.8
onnxruntime=1.7.2
scikit-learn
pandas
numpy
h5py
pip
git+https://github.com/joefutrelle/pyifcb.git
